+++ index.js
@@ -185,7 +185,7 @@ function generateRewritePrefix (prefix = '', opts) {
   return rewritePrefix
 }
 
-async function httpProxy (fastify, opts) {
+async function fastifyHttpProxy (fastify, opts) {
   if (!opts.upstream && !(opts.upstream === '' && opts.replyOptions && typeof opts.replyOptions.getUpstream === 'function')) {
     throw new Error('upstream must be specified')
   }
@@ -261,11 +261,11 @@ async function httpProxy (fastify, opts) {
   }
 }
 
-httpProxy[Symbol.for('plugin-meta')] = {
+fastifyHttpProxy[Symbol.for('plugin-meta')] = {
   fastify: '4.x',
   name: '@fastify/http-proxy'
 }
 
-module.exports = httpProxy
-module.exports.default = httpProxy
-module.exports.fastifyHttpProxy = httpProxy
+module.exports = fastifyHttpProxy
+module.exports.default = fastifyHttpProxy
+module.exports.fastifyHttpProxy = fastifyHttpProxy
